+++ perfkitbenchmarker/pkb.py
@@ -138,7 +138,13 @@ flags.DEFINE_enum(
     'Type for all scratch disks. The default is standard')
 flags.DEFINE_integer('scratch_disk_iops', 1500,
                      'IOPS for Provisioned IOPS (SSD) volumes in AWS.')
-
+flags.DEFINE_integer('num_striped_disks', 1,
+                     'The number of disks to stripe together to form one '
+                     '"logical" scratch disk. This defaults to 1 '
+                     '(except with local disks), which means no striping. '
+                     'When using local disks, they default to striping '
+                     'all disks together.',
+                     lower_bound=1)
 
 MAX_RUN_URI_LENGTH = 10